refactor(stop-search-index): move to domain/transit
`buildSearchIndexEntry` and `filterStopsByQuery` are GTFS-specific
search logic — they read `Stop.stop_name` / `stop_names`, sort by
`ja-Hrkt` locale order, and assume the multi-language Stop shape from
the transit data. Per the project's placement rule (DEVELOPMENT.md:
"アプリ固有の判断を含むなら utils ではなく domain"), they belong in
`src/domain/transit/` alongside `distance.ts`, `get-stop-display-names.ts`,
`stop-history.ts`, etc., not in the generic `src/utils/` bucket.

Pure file move + import-path updates — no behaviour change.

- src/utils/stop-search-index.ts → src/domain/transit/stop-search-index.ts
- src/utils/stop-search-index.test.ts → src/domain/transit/__tests__/stop-search-index.test.ts
- src/components/dialog/stop-search-dialog.tsx — import path
- src/hooks/use-stop-search-index.ts — import path
- src/domain/transit/stop-search-index.ts itself now imports
  `katakanaToHiragana` via the `@/utils/kana-normalize` alias instead
  of the old sibling-relative path
